23683 MessageProducer should be interface MessageProducer is now interface
